+++ src/main/assets/changelog.md
@@ -1,4 +1,5 @@
-### v5.5.5 (93)
+### v5.6.0
 - Remove JNI requirement, Magisk Manager is now pure Java
-- Update the method to handle global su database (v15+), should fix root request not saving issue
-on many devices
+- Update the method of handling su database, may fix the issue that root requests won't save
+- Add the option to restore Magisk Manager after repackaging with random package name
+- Massive under-the-hood optimizations